Keep the frame address comparison at 64 bits
The merge resolution reverted frameAddressHigh to uint32. The value is a
pointer shifted right by 16, which fits in 32 bits only because current x86-64
user-space addresses are 47-bit. Keep it uint64 so the comparison does not
silently truncate if that ever stops holding.

+++ engine/server/library/serverScript/src/shared/JavaLibrary.cpp
@@ -811,7 +811,7 @@ void JavaLibrary::fatalHandler(int signum)
 		void *crashAddress2c = nullptr;
 		void *frameAddressA = nullptr;
 		void *frameAddressB = nullptr;
-		uint32 frameAddressHigh = (reinterpret_cast<uint64>(frameAddress) >> 16);
+		uint64 frameAddressHigh = (reinterpret_cast<uint64>(frameAddress) >> 16);
 		crashAddress2a = __builtin_return_address(0);
 // Suppress Wframe-address for these lines - we could crash the program calling __builtin_return_address and frame_address
 // with non-zero values.  However, we likely don't care as we're crashing at this point anyway due to bad Java.
